The problem of distinguishing prime numbers from composite numbers and of resolving the latter into their prime factors is known to be one of the most important and useful in arithmetic

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,9 +5,9 @@
   * Return: Returns 1 if True, 0 otherwise
   */
 
-int is_prime(unsigned long long int num)
+int is_prime(unsigned long int num)
 {
-	unsigned long long int i;
+	unsigned long int i;
 
 	for (i = num - 1; i >= 2; i--)
 		if (num % i == 0)
@@ -23,12 +23,12 @@ int is_prime(unsigned long long int num)
 
 int main(void)
 {
-	unsigned long long int num = 612852475143, i;
+	unsigned long int num = 612852475143, i;
 
 	for (i = num - 1; i >= 2; i--)
 		if (is_prime(i) == 1 && num % i == 0)
 		{
-			printf("%llu\n", i);
+			printf("%lu\n", i);
 			break;
 		}
